Created Failed Test Notifications
Built a menu screen to appear when a wrong answer is selected during the test, and another similar menus screen if the user fails the test by running out of time. NOTE: When these events occur, the intent is to dynamically change the background colors of the header and footer, as well as the content of the Button.