Fixes Bug of Next Button
Previously, the correctAnswer variable was holding only the first correct answer of the list of three on the DOM. This was resulting in a Next Button that would move the next question even if the "correct answer" for the active question was not selected. Although the correct answer from the previous question was no longer visible, it was remaining checked invisibly and so allowing the next button to function. This fixes that bug.