Tries to Create a Switch statement for Next Button
This is a draft branch. Was working to create a switch statement to change the wrong answers being checked based on the current active panel. Wanted to dynamically change which wrong answers were being selected to be checked against based on the active panel. Was unable to achieve this and I am temporarily abandoning this approach until a future time when I've done more research. The mess of code is my active trial and error state.