Adds Conditions for Menus Display
Created functions for logic in deciding which menu should be display based upon the interaction of the user. Now, the "Failed: Wrong Answer" menu displays if a wrong answer has been selected, the "Failed: Timed Out" menu displays if the timer runs out (NOTE: timer API is not yet implemented), and the "Nothing Was Selected" notice displays if the user tries to advance to the next question without selecting an answer.