Temporarily Fixes Quiz Footer Elements Bug
Extra JavaScript was created to change the display value of the Quiz Footer Elements to "none" upon page load, and then back to "flex" when the start button is selected. The original bug is believed to have been caused by a specificity issue, but has not been resolved. This fix makes the program functional again, but may not be the best way to resolve the original issue.