Removes Failed Function Switches Approach
The deleted code was used as a first approach for encapsulating the code to remove/replace various Footer Elements Groups. In this approach, a bug would occur in which the makeInvisible case was successful, but the makeVisible case would not perform its function. This approach was abandoned and replaced with another successful approach, but kept for sake of archiving attempts.