Creates Code to Search For Bug
Discovered a bug in which, after selecting the Main Menu Button, the user would not be able to progress through to the 3rd question because regardless of what they would select (on the second time through the quiz) the "Nothing Was Selected" page would always show up. It seems although, after returning to the start menu, something happens in which the function checking for wrong answers selected fails to work (sometimes on 1st question, sometimes on second question, and sometimes on 3rd question).