Fixes Bug that Caused Wrong-Answer Check to Fail
The bug, which caused the "Nothing Was Selected" notice to display even when something was selected, originated in the next.Button Event Listener. This was the first Event Listener to be created, and was originally designed to act as the designator determining which fail notice/menu would appear based on the state of what was selected. Further development moved away from this approach, but the skeleton of that design remained. That skeleton was flawed; the else if statement was designed to check if any wrong answers were selected, but in practice, its code was just checking if the wrongAnswerChecked() function existed, which it did. It seems as though the function would run in the else-if statement, but the code dependent on if it existed never ran. Removing the else-if statement and moving the function to the else statement resolved this bug.