Corrects Radio Button Word Wrap Issue
In the case where the total length of words was more than the view width, the words would wrap to the next line in a manner that messed with the design. They would create a new line below the radio input itself. Wrapping the input and label element together in a div with display flex resolved this issue.