Fixes Timer Bug
The timer previously created had no means for stopping. The bug would occur after the time ran out, in which case the time would never stop running out and any action would result in the immediate display of the Timed Out notice. The fix, was to include "clearInterval(time)" after the countdown ended.