Fixes Timer Bug After User Victory
The timer failed to reset if the quiz was restarted after the user completes the quiz successfully. This was resolved by adding in some log to check if the test was completed and if so, to conditionally alter the functionality of the timer. Main Menu Button would also check the condition of the completion of the test to determine if the timer needed to be reset.