Adds Quit Button & Functionality
Created a Quit Button to work alongside the Main Menu Button.  The functionality added in the Event Listener for this new button was copied almost exactly from the Main Menu Button Event Listener. The intent is that the Quit Button will take over some of the Main Menu Button's previous controls. The Quit Button will now return you to the main menu, while the main menu button will be intended to take the user directly back to question 1 for a retaking of the quiz.