Fixes Quit Button Bug
Because of the new state created by the "Quitting Pug" Panel, various functionalities of the Quit Button Event Listener broke. This caused a vast array of bugs at different states of the quiz and under varying conditions. This restructuring solves this issue but evaluating first if the Quitting Pug Panel is active, then evaluating if the "Winner Screen" Panel is active, and finally executing its code in an if/else if/ else statement.