Fixes Quit Button Bug
The quit button is supposed to reset the game stat for the quiz, however, the timer failed to reset from this event. The problem was a matter of scope. The fix was to create a new Event Listener on the Quit Button from within the createTimer() function, and then remove it after its use. Due to the matter of the createTimer() function's scope, this Event Listener will have to be created and destroyed within every call of the function.